+++ packages/components/src/components/general/index.ts
@@ -1,0 +1,2 @@
+export { ThemeProvider, useTheme } from "./theme-provider.js";
+export { BrowserOnlyBanner } from "./browser-only-banner.js";